src/requirements/tests.txt
-r base.txt

flake8==3.5.0
pytest==3.6.1
pytest-cov==2.5.1
pytest-flask==0.10.0
pytest-runner==4.2
pymongo==3.6.1
flask-mongoengine==0.9.5
mongoengine==0.15.0